Module google_prediction1d6::api

Structs

Analyze
There is no detailed description.

AnalyzeDataDescription
Description of the data the model was trained on.

AnalyzeDataDescriptionFeatures
Description of the input features in the data set.

AnalyzeDataDescriptionFeaturesCategorical
Description of the categorical values of this feature.

AnalyzeDataDescriptionFeaturesCategoricalValues
List of all the categories for this feature in the data set.

AnalyzeDataDescriptionFeaturesNumeric
Description of the numeric values of this feature.

AnalyzeDataDescriptionFeaturesText
Description of multiple-word text values of this feature.

AnalyzeDataDescriptionOutputFeature
Description of the output value or label.

AnalyzeDataDescriptionOutputFeatureNumeric
Description of the output values in the data set.

AnalyzeDataDescriptionOutputFeatureText
Description of the output labels in the data set.

AnalyzeModelDescription
Description of the model.

HostedmodelMethods
A builder providing access to all methods supported on hostedmodel resources. It is not used directly, but through the Prediction hub.

HostedmodelPredictCall
Submit input and request an output against a hosted model.

Input
There is no detailed description.

InputInput
Input to the model for a prediction.

Insert
There is no detailed description.

Insert2
There is no detailed description.

Insert2ModelInfo
Model metadata.

InsertTrainingInstances
Instances to train model on.

List
There is no detailed description.

Output
There is no detailed description.

OutputOutputMulti
A list of class labels with their estimated probabilities (Categorical models only).

Prediction
Central instance to access all Prediction related resource activities

TrainedmodelAnalyzeCall
Get analysis of the model and the data the model was trained on.

TrainedmodelDeleteCall
Delete a trained model.

TrainedmodelGetCall
Check training status of your model.

TrainedmodelInsertCall
Train a Prediction API model.

TrainedmodelListCall
List available models.

TrainedmodelMethods
A builder providing access to all methods supported on trainedmodel resources. It is not used directly, but through the Prediction hub.

TrainedmodelPredictCall
Submit model id and request a prediction.

TrainedmodelUpdateCall
Add new data to a trained model.

Update
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.