Struct google_prediction1d6::Insert

pub struct Insert {
    pub storage_data_location: Option<String>,
    pub model_type: Option<String>,
    pub storage_pmml_model_location: Option<String>,
    pub source_model: Option<String>,
    pub storage_pmml_location: Option<String>,
    pub training_instances: Option<Vec<InsertTrainingInstances>>,
    pub id: Option<String>,
    pub utility: Option<Vec<HashMap<String, f64>>>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert trainedmodels (request)

Fields

storage_data_location

Google storage location of the training data file.

model_type

Type of predictive model (classification or regression).

storage_pmml_model_location

Google storage location of the pmml model file.

source_model

The Id of the model to be copied over.

storage_pmml_location

Google storage location of the preprocessing pmml file.

training_instances

Instances to train model on.

id

The unique name for the predictive model.

utility

A class weighting function, which allows the importance weights for class labels to be specified (Categorical models only).