Struct google_prediction1d6::Insert2

pub struct Insert2 {
    pub kind: Option<String>,
    pub created: Option<String>,
    pub storage_pmml_location: Option<String>,
    pub storage_pmml_model_location: Option<String>,
    pub model_type: Option<String>,
    pub storage_data_location: Option<String>,
    pub self_link: Option<String>,
    pub model_info: Option<Insert2ModelInfo>,
    pub training_complete: Option<String>,
    pub id: Option<String>,
    pub training_status: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert trainedmodels (response)

get trainedmodels (response)

update trainedmodels (response)

Fields
kind: Option<String>

What kind of resource this is.

created: Option<String>

Insert time of the model (as a RFC 3339 timestamp).

storage_pmml_location: Option<String>

Google storage location of the preprocessing pmml file.

storage_pmml_model_location: Option<String>

Google storage location of the pmml model file.

model_type: Option<String>

Type of predictive model (CLASSIFICATION or REGRESSION).

storage_data_location: Option<String>

Google storage location of the training data file.

self_link: Option<String>

A URL to re-request this resource.

model_info: Option<Insert2ModelInfo>

Model metadata.

training_complete: Option<String>

Training completion time (as a RFC 3339 timestamp).

id: Option<String>

The unique name for the predictive model.

training_status: Option<String>

The current status of the training job. This can be one of following: RUNNING; DONE; ERROR; ERROR: TRAINING JOB NOT FOUND