Struct google_prediction1d6::Insert2ModelInfo

pub struct Insert2ModelInfo {
    pub number_instances: Option<String>,
    pub class_weighted_accuracy: Option<String>,
    pub number_labels: Option<String>,
    pub classification_accuracy: Option<String>,
    pub mean_squared_error: Option<String>,
    pub model_type: Option<String>,
}

Model metadata.

This type is not used in any activity, and only used as part of another schema.

Fields

number_instances

Number of valid data instances used in the trained model.

class_weighted_accuracy

Estimated accuracy of model taking utility weights into account (Categorical models only).

number_labels

Number of class labels in the trained model (Categorical models only).

classification_accuracy

A number between 0.0 and 1.0, where 1.0 is 100% accurate. This is an estimate, based on the amount and quality of the training data, of the estimated prediction accuracy. You can use this is a guide to decide whether the results are accurate enough for your needs. This estimate will be more reliable if your real input data is similar to your training data (Categorical models only).

mean_squared_error

An estimated mean squared error. The can be used to measure the quality of the predicted model (Regression models only).

model_type

Type of predictive model (CLASSIFICATION or REGRESSION).