Trait google_prediction1d6::NestedType

pub trait NestedType { }

Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.